# docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: Prospect API
  version: 0.1.0
  description: Evidence-first financial document intelligence API.
servers:
  - url: /api/v1
security:
  - sessionToken: []

# Every non-2xx response body is an Error. A document owned by another user returns the
# same 404 as a nonexistent one.
paths:
  /sessions:
    post:
      summary: Start an anonymous single-user-mode session (replaced by real auth later)
      description: >
        Token format v1.<session_id>.<issued_at>.<expires_at>.<hmac>. Rejected (401) when
        malformed, badly signed, expired, revoked or unknown. Rate limited per client IP.
        See docs/SECURITY.md.
      security: []
      responses:
        "201":
          description: Signed, expiring session token for the Authorization header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionToken"
        "429":
          description: Rate limited; Retry-After gives the seconds to wait
  /sessions/refresh:
    post:
      summary: Exchange a valid token for a fresh one (same user); revokes the old session
      responses:
        "200":
          description: New session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionToken"
        "401":
          description: Missing, malformed, expired, revoked or unknown session
        "429":
          description: Rate limited (Retry-After header)
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
...
  schemas:
...
    SessionToken:
      type: object
      required: [token, expires_at]
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
...